+++ views/default/amap_maps_api/location_map.js
@@ -1,7 +1,10 @@
 define(function (require) {
     var elgg = require('elgg');
-	var $ = require('jquery');
+    var $ = require('jquery');
     require('amap_ma_googleapis_js')
+    
+    
+    var map_settings = require("amap_maps_api/settings");
     
     $( document ).ready(function() {
            
@@ -9,6 +12,7 @@ define(function (require) {
         var entity_lat = $('#entity_lat').text();
         var entity_lon = $('#entity_lon').text();
         var zoom = parseInt($('#map_zoom').text());
+        var map_center = $('#map_center').text();
         var marker = $('#entity_marker').text();
         var gm = google.maps;   
 
@@ -19,24 +23,43 @@ define(function (require) {
 
         infowindow = new google.maps.InfoWindow();
         var myLatlng = new google.maps.LatLng(entity_lat,entity_lon);
+        var cLatlng = myLatlng;
+        if (map_center) {
+            var center_arr = map_center.split(",");
+            cLatlng = new google.maps.LatLng(center_arr[0],center_arr[1]);
+        }
+        
         var mapOptions = {
             zoom: zoom,
-            center: myLatlng,
+            center: cLatlng,
             mapTypeId: gm.MapTypeId.ROADMAP
         }
         
-        var map = new gm.Map(document.getElementById("map"), mapOptions);
+        var location_map = new gm.Map(document.getElementById("location_map"), mapOptions);
+        location_map.setMapTypeId(map_settings['default_layer']);
+        
+        location_map.mapTypes.set("OSM", new google.maps.ImageMapType({
+            getTileUrl: function(coord, zoom) {
+                // See above example if you need smooth wrapping at 180th meridian
+                return map_settings['osm_base_layer'] + zoom + "/" + coord.x + "/" + coord.y + ".png";
+            },
+            tileSize: new google.maps.Size(256, 256),
+            name: "OpenStreetMap",
+            maxZoom: 18
+        })); 
+        
         var marker = new google.maps.Marker({
-            map: map,
+            map: location_map,
             position: myLatlng,
             title: entity_title,
             icon: marker,
         });                
         google.maps.event.addListener(marker, "click", function() {
           infowindow.setContent(entity_title);
-          infowindow.open(map, this);
+          infowindow.open(location_map, this);
         });  
-        map.setCenter(marker.getPosition());        
+        
+        location_map.setCenter(cLatlng);
         
         return true;		
     });